Custom weight from int to float.

+++ mystery_shopping/questionnaires/models.py
@@ -276,7 +276,7 @@ class CustomWeight(models.Model):
 
     # Attributes
     name = models.CharField(max_length=200)
-    weight = models.IntegerField(default=0)
+    weight = models.FloatField(default=0)
 
     objects = models.Manager.from_queryset(CustomWeightQuerySet)()
